Move vertex structures to `Batches` namespace
