+++ osu.Framework/Graphics/Batches/ParticleVertex2D.cs
@@ -1,18 +1,19 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
 #nullable disable
 
 using System;
 using System.Runtime.InteropServices;
+using osu.Framework.Graphics.OpenGL.Vertices;
 using osuTK;
 using osuTK.Graphics;
 using osuTK.Graphics.ES30;
 
-namespace osu.Framework.Graphics.OpenGL.Vertices
+namespace osu.Framework.Graphics.Batches
 {
     [StructLayout(LayoutKind.Sequential)]
-    public struct TimedTexturedVertex2D : IEquatable<TimedTexturedVertex2D>, IVertex
+    public struct ParticleVertex2D : IEquatable<ParticleVertex2D>, IVertex
     {
         [VertexMember(2, VertexAttribPointerType.Float)]
         public Vector2 Position;
@@ -26,6 +27,9 @@ namespace osu.Framework.Graphics.OpenGL.Vertices
         [VertexMember(1, VertexAttribPointerType.Float)]
         public float Time;
 
-        public readonly bool Equals(TimedTexturedVertex2D other) => Position.Equals(other.Position) && TexturePosition.Equals(other.TexturePosition) && Colour.Equals(other.Colour) && Time.Equals(other.Time);
+        [VertexMember(2, VertexAttribPointerType.Float)]
+        public Vector2 Direction;
+
+        public readonly bool Equals(ParticleVertex2D other) => Position.Equals(other.Position) && TexturePosition.Equals(other.TexturePosition) && Colour.Equals(other.Colour) && Time.Equals(other.Time) && Direction.Equals(other.Direction);
     }
 }
