+++ osu.Framework/Graphics/Batches/VertexMemberAttribute.cs
@@ -6,7 +6,7 @@
 using System;
 using osuTK.Graphics.ES30;
 
-namespace osu.Framework.Graphics.OpenGL.Vertices
+namespace osu.Framework.Graphics.Batches
 {
     [AttributeUsage(AttributeTargets.Field)]
     public class VertexMemberAttribute : Attribute
