+++ osu.Framework/Graphics/UserInterface/CircularProgressDrawNode.cs
@@ -10,7 +10,7 @@ using System;
 using osu.Framework.Graphics.Primitives;
 using osuTK.Graphics;
 using osu.Framework.Extensions.MatrixExtensions;
-using osu.Framework.Graphics.OpenGL.Vertices;
+using osu.Framework.Graphics.Batches;
 using osu.Framework.Graphics.Rendering;
 
 namespace osu.Framework.Graphics.UserInterface
