+++ osu.Framework/osu.Framework.csproj
@@ -55,4 +55,7 @@
          See https://github.com/NuGet/Home/issues/4514 and https://github.com/dotnet/sdk/issues/765 . -->
     <PackageReference Include="OpenTabletDriver" Version="0.6.0.2" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Graphics\Vertices" />
+  </ItemGroup>
 </Project>
